Outsource calendar types
